djang-docker-drf-nodeblog/requirements.txt
Django==2.2.5
django-filter==2.2.0
djangorestframework==3.10.3
gunicorn==19.6.0
Markdown==3.1.1
pytz==2019.2
sqlparse==0.3.0

